Move PrincipalType to presto-spi
