+++ presto-hive/src/main/java/com/facebook/presto/hive/metastore/file/DatabaseMetadata.java
@@ -14,7 +14,7 @@
 package com.facebook.presto.hive.metastore.file;
 
 import com.facebook.presto.hive.metastore.Database;
-import com.facebook.presto.hive.metastore.PrincipalType;
+import com.facebook.presto.spi.security.PrincipalType;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.collect.ImmutableMap;
