+++ presto-hive/src/test/java/com/facebook/presto/hive/HiveBenchmarkQueryRunner.java
@@ -16,8 +16,8 @@ package com.facebook.presto.hive;
 import com.facebook.presto.Session;
 import com.facebook.presto.benchmark.BenchmarkSuite;
 import com.facebook.presto.hive.metastore.Database;
-import com.facebook.presto.hive.metastore.PrincipalType;
 import com.facebook.presto.hive.metastore.TestingHiveMetastore;
+import com.facebook.presto.spi.security.PrincipalType;
 import com.facebook.presto.testing.LocalQueryRunner;
 import com.facebook.presto.tpch.TpchConnectorFactory;
 import com.google.common.collect.ImmutableMap;
