+++ presto-spi/src/main/java/com/facebook/presto/spi/security/PrincipalType.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive.metastore;
+package com.facebook.presto.spi.security;
 
 public enum PrincipalType
 {
